backend/requirements.txt
fastapi>=0.135.1
uvicorn>=0.30.0
groq==1.0.0
python-dotenv==1.2.2
pydantic==2.12.5
httpx==0.28.1
anyio==4.12.1
slowapi>=0.1.9
